Headers will be passed as plain object
Earlier headers was passed as a special read-only object as defined by the `fetch` API.

The entries were returned as a iterator. And one had to use a special method `get` to retrieve the value for a known key in the header.

Now it will be passed as a plain object